separate host implementation from library
