+++ crates/voxea_vst/src/vst/audio_processor.rs
@@ -1097,51 +1097,3 @@ pub trait IParameterChanges: FUnknown {
     fn get_parameter_data(&mut self) -> *mut c_void;
     fn add_parameter_data(&mut self, id: *const c_char, index: *mut i32) -> *mut c_void;
 }
-
-#[repr(C)]
-pub struct HostParameterChanges {
-    vtable: &'static [*const (); 6],
-}
-
-impl HostParameterChanges {
-    pub fn new() -> Self {
-        Self {
-            vtable: &[
-                <Self as FUnknown_HostImpl>::query_interface as *const (),
-                <Self as FUnknown_HostImpl>::add_ref as *const (),
-                <Self as FUnknown_HostImpl>::release as *const (),
-                <Self as IParameterChanges_HostImpl>::get_parameter_count as *const (),
-                <Self as IParameterChanges_HostImpl>::get_parameter_data as *const (),
-                <Self as IParameterChanges_HostImpl>::add_parameter_data as *const (),
-            ],
-        }
-    }
-}
-
-impl Interface for HostParameterChanges {
-    type VTable = [*const (); 6];
-    const iid: FUID = [0; 16];
-
-    fn vtable(&self) -> &'static Self::VTable {
-        self.vtable
-    }
-}
-
-impl FUnknown_HostImpl for HostParameterChanges {}
-
-impl IParameterChanges_HostImpl for HostParameterChanges {
-    unsafe fn get_parameter_count(&mut self) -> i32 {
-        warn!("get_parameter_count");
-        0
-    }
-
-    unsafe fn get_parameter_data(&mut self) -> *mut c_void {
-        warn!("get_parameter_data");
-        std::ptr::null_mut()
-    }
-
-    unsafe fn add_parameter_data(&mut self, id: *const c_char, index: *mut i32) -> *mut c_void {
-        warn!("add_parameter_data");
-        std::ptr::null_mut()
-    }
-}
